Fix i64 overflow at k ≥ 63 (correctness, slower at ε≤1e-7)
The integer Lenstra path had silent integer overflow at k ≥ 63:
- target_norm = 1i64 << k → i64::MIN (k=63) or UB (k≥64)
- threshold_xy = (1u128 << 2k) overflowed at k ≥ 64
- ‖x‖² and bilinear_b(x) computed in i64 → wrap on x_i ≈ 2^41 products
- uv_to_xy used (1i64 << k) for the alignment-vector scale

The k > 62 guard at synthesizer.rs:430 was a workaround; it silently
skipped all prefixes with k_inner > 62. At ε ≤ 1e-7 (lde ≥ 60+), this
means many valid prefixes never got searched. Solutions were still
found (via different t_prime values where k_inner ≤ 62) but the search
was incomplete.

Changes:
- target_norm widened to i128
- threshold_xy computed via 2.0_f64.powi(2*k) — exact in f64 for k ≤ 1023
- bilinear_b returns i128 (computed in i128 throughout)
- SE callback ‖x‖² check has fast i64 path for k ≤ 62, i128 path for k > 62
  (the SE callback runs millions of times; keeping i64 when safe avoids
  ~3-5× per-op cost on aarch64)
- uv_to_xy uses 2.0_f64.powf(k/2 - 1)
- k > 62 guard bumped to k > 110 (i128 headroom for sums + bilinear_b)

Performance impact:
- ε=1e-5: unchanged (k stays ≤ 62, fast path engaged)
- ε=1e-7 Rz(pi/7): 6900ms → 13400ms (~2× slower) — now correctly
  searches prefixes that were silently skipped before
- ε=1e-8 Rz(0.30): 117s → 230s (~2× slower) — same reason

The slowdown at deep ε is the cost of a correct, complete search. The
old fast behavior was an artifact of silent i64 overflow rejecting
most candidates at high-k prefixes; correctness > speed for this
fix. Future optimization: improve i128 hot-path performance, or
prefix-level pruning to skip provably-empty prefixes.

All 131 lib tests pass.

+++ cyclosynth-rs/src/synthesis/lenstra_heavy.rs
@@ -952,9 +952,12 @@ where
 // ─── Top-level phase1_lenstra ─────────────────────────────────────────────────
 
 #[inline]
-pub(crate) fn bilinear_b(x: &[i64; 8]) -> i64 {
-    let (a1, b1, c1, d1) = (x[0], x[1], x[2], x[3]);
-    let (a2, b2, c2, d2) = (x[4], x[5], x[6], x[7]);
+pub(crate) fn bilinear_b(x: &[i64; 8]) -> i128 {
+    // Returns i128 because at deep ε (k ≥ 63), individual x_i can reach ~2^41
+    // and products x_i·x_j hit ~2^82 — overflows i64 silently. Computed in
+    // i128 throughout so the == 0 check is correct at all k.
+    let (a1, b1, c1, d1) = (x[0] as i128, x[1] as i128, x[2] as i128, x[3] as i128);
+    let (a2, b2, c2, d2) = (x[4] as i128, x[5] as i128, x[6] as i128, x[7] as i128);
     a1 * b1 - a1 * d1 + b1 * c1 + c1 * d1 + a2 * b2 - a2 * d2 + b2 * c2 + c2 * d2
 }
